Iterate over log events & post them individually
CloudWatch often bundles multiple log events together in the event input
archive that is passed in the lambda's event argument. Since the lambda
was only doing a single post to the collection, that post would in many
cases actually contain multiple logs which would then show up as a
single concatenated log in the collection resulting in incorrect metrics.

This change iterates over the log events in the archive and posts them
individually to the collection.