Allow an optional custom Lambda function name
Add an optional parameter to specify an explicit name for the log
processor Lambda function.

In most cases, using the default CloudFormation-generated name is the
right approach. For some cases though, there is a benefit to having a
predictable function name that is reasonably static. One example is
migrating from a legacy standalone (non-DLQ/CloudFormation) Sumo Lambda
setup, which has a number of active subscriptions maintained by many
stakeholders. In a situation like that, a known function name can help
ease the transition.